fix(api): saindo de decimal para double no location migration

+++ api/database/migrations/2025_01_31_114835_create_locations_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('locations', function (Blueprint $table) {
             $table->id();
-            $table->decimal('latitude');
-            $table->decimal('longitude');
+            $table->double('latitude');
+            $table->double('longitude');
             $table->string('photo');
             $table->timestamps();
         });
